[rollout] refactor: move MiniMax H3 agent-loop helpers into the pipeline package
messages_to_text and MINIMAX_H3_TOKEN_ID_NATIVE_KEY are only used by the
MiniMax H3 agent loop and pipeline; host them in
pipelines/minimax_h3_diffusion_nft/common.py and restore
verl_omni/agent_loop/utils.py to upstream.
